+++ java2d/src/test/scala/doodle/java2d/Base64Spec.scala
@@ -1,38 +1,43 @@
 package doodle
 package java2d
 
+import doodle.core.Base64
 import doodle.effect.Writer._
 import doodle.syntax._
 import minitest._
 
 object Base64Spec extends SimpleTestSuite {
-  val image = circle[Algebra,Drawing](20.0)
+  def base64Distance[A](b1: Base64[A], b2: Base64[A]): Double = {
+    import java.util.{Base64 => JBase64}
+    val d1 = JBase64.getDecoder().decode(b1.value)
+    val d2 = JBase64.getDecoder().decode(b2.value)
 
-  test("write should work with png") {
-    val (_, string1) = image.base64[Png]()
-    val (_, string2) = image.base64[Png]()
-
-    assertEquals(string1, string2)
+    d1.zip(d2).foldLeft(0.0) { (accum, elts) =>
+      val (byte1, byte2) = elts
+      accum + Math.abs(byte1 - byte2)
+    }
   }
 
-  test("write should work with gif") {
-    val (_, string1) = image.base64[Gif]()
-    val (_, string2) = image.base64[Gif]()
+  val image = circle[Algebra, Drawing](20.0)
 
-    assertEquals(string1, string2)
+  test("base64 should work with png") {
+    val (_, b1) = image.base64[Png]()
+    val (_, b2) = image.base64[Png]()
+
+    assert(base64Distance(b1, b2) <= (b1.value.length * 2))
   }
 
-  test("write should work with pdf") {
-    val (_, string1) = image.base64[Pdf]()
-    val (_, string2) = image.base64[Pdf]()
+  test("base64 should work with gif") {
+    val (_, b1) = image.base64[Gif]()
+    val (_, b2) = image.base64[Gif]()
 
-    assertEquals(string1, string2)
+    assert(base64Distance(b1, b2) <= (b1.value.length * 2))
   }
 
-  test("write should work with jpg") {
-    val (_, string1) = image.base64[Jpg]()
-    val (_, string2) = image.base64[Jpg]()
+  test("base64 should work with jpg") {
+    val (_, b1) = image.base64[Jpg]()
+    val (_, b2) = image.base64[Jpg]()
 
-    assertEquals(string1, string2)
+    assert(base64Distance(b1, b2) <= (b1.value.length * 2))
   }
 }